test/fixtures/wpt/html/webappapis/timers/settimeout-cross-realm-callback-report-exception.html

<!doctype html>
<meta charset=utf-8>
<title>window.setTimeout() reports the exception from its callback in the callback's global object</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<iframe></iframe>
<iframe></iframe>
<iframe></iframe>
<script>
setup({ allow_uncaught_exception: true });

const onerrorCalls = [];
window.onerror = () => { onerrorCalls.push("top"); };
frames[0].onerror = () => { onerrorCalls.push("frame0"); };
frames[1].onerror = () => { onerrorCalls.push("frame1"); };
frames[2].onerror = () => { onerrorCalls.push("frame2"); };

async_test(t => {
  window.onload = t.step_func(() => {
    frames[0].setTimeout(new frames[1].Function(`throw new parent.frames[2].Error("PASS");`), 4);

    t.step_timeout(() => {
      assert_array_equals(onerrorCalls, ["frame1"]);
      t.done();
    }, 8);
  });
});
</script>
