LayoutTests/imported/w3c/web-platform-tests/html/webappapis/timers/settimeout-cross-realm-callback-report-exception-expected.txt

CONSOLE MESSAGE: Error: PASS


PASS window.setTimeout() reports the exception from its callback in the callback's global object
